requirements.txt
numpy
opencv-python
mss
ultralytics
torch
PyYAML
pydirectinput
pynput
